requirements.txt
faiss-cpu
numpy
tqdm
torch
transformers
Pillow
